Reorganize modules and update docstring
