+++ dataikuapi/govern/blueprint.py
@@ -1,7 +1,7 @@
 class GovernBlueprint(object):
     """
-    A handle to read a blueprint on the Govern instance. If you wish to edit blueprints or the blueprint versions, use
-    :meth:`dataikuapi.govern_client.get_admin_blueprint_designer()`.
+    A handle to read a blueprint on the Govern instance. If you wish to edit blueprints or the blueprint versions, use 
+    the blueprint designer object :class:`dataikuapi.admin_blueprint_designer.GovernAdminBlueprintDesigner()`.
     Do not create this directly, use :meth:`dataikuapi.govern_client.GovernClient.get_blueprint()`
     """
 
@@ -22,8 +22,8 @@ class GovernBlueprint(object):
         """
         Return the definition of the blueprint as an object.
 
-        :returns: the blueprint definition as an object.
-        :rtype: :class:`dataikuapi.govern.models.GovernBlueprintDefinition`
+        :returns: The blueprint definition as an object.
+        :rtype: :class:`dataikuapi.govern.blueprint.GovernBlueprintDefinition`
         """
         definition = self.client._perform_json(
             "GET", "/blueprint/%s" % self.blueprint_id).get("definition")
@@ -33,10 +33,10 @@ class GovernBlueprint(object):
         """
         Lists versions of this blueprint.
 
-        :param boolean as_objects: (Optional) if True, returns a list of :class:`GovernAPIBlueprintDesignerBlueprintVersion`, else
-         returns a list of dict. Each dict contains a field "id" indicating the identifier of this version
-        :returns: a list - see as_objects for more information
-        :rtype: list of :class: `dataikuapi.govern.models.GovernBlueprintVersion` or list of dict
+        :param boolean as_objects: (Optional) if True, returns a list of :class:`GovernAPIBlueprintDesignerBlueprintVersion`
+        , else returns a list of dict. Each dict contains a field "id" indicating the identifier of this version
+        :returns: The list of the versions
+        :rtype: list of :class:`dataikuapi.govern.blueprint.GovernBlueprintVersion` or list of dict
         """
         versions = self.client._perform_json(
             "GET", "/blueprint/%s/versions" % self.id)
@@ -51,7 +51,7 @@ class GovernBlueprint(object):
         Returns a handle to interact with a blueprint version
 
         :param str version_id: id of the version
-        :rtype: :class: `dataikuapi.govern.models.GovernBlueprintVersion`
+        :rtype: :class:`dataikuapi.govern.models.GovernBlueprintVersion`
         """
         return GovernBlueprintVersion(self.client, self.blueprint_id, version_id)
 
@@ -59,7 +59,7 @@ class GovernBlueprint(object):
 class GovernBlueprintDefinition(object):
     """
     A handle of the definition of a blueprint
-    Do not create this class directly, instead use :meth:`dataikuapi.govern.models.GovernBlueprint.get_definition`
+    Do not create this class directly, instead use :meth:`dataikuapi.govern.blueprint.GovernBlueprint.get_definition()`
     """
 
     def __init__(self, client, blueprint_id, definition):
@@ -71,7 +71,7 @@ class GovernBlueprintDefinition(object):
         """
         Get raw definition of a blueprint
 
-        :return: the raw definition of blueprint, as a dict.
+        :return: The raw definition of blueprint, as a dict.
         :rtype: dict
         """
         return self.definition
@@ -81,7 +81,7 @@ class GovernBlueprintDefinition(object):
         """
         Return the blueprint id.
 
-        :return: the blueprint id
+        :return: The blueprint id
         :rtype: str
         """
         return self.blueprint_id
@@ -91,7 +91,7 @@ class GovernBlueprintDefinition(object):
         """
         Return the blueprint name.
 
-        :return: the blueprint name
+        :return: The blueprint name
         :rtype: str
         """
         return self.definition.get("name")
@@ -101,7 +101,7 @@ class GovernBlueprintDefinition(object):
         """
         Return the blueprint icon.
 
-        :return: the blueprint icon
+        :return: The blueprint icon
         :rtype: str
         """
         return self.definition.get("icon")
@@ -111,7 +111,7 @@ class GovernBlueprintDefinition(object):
         """
         Return the blueprint color.
 
-        :return: the blueprint color
+        :return: The blueprint color
         :rtype: str
         """
         return self.definition.get("color")
@@ -121,7 +121,7 @@ class GovernBlueprintDefinition(object):
         """
         Return the blueprint background color.
 
-        :return: the blueprint background color
+        :return: The blueprint background color
         :rtype: str
         """
         return self.definition.get("backgroundColor")
@@ -130,7 +130,7 @@ class GovernBlueprintDefinition(object):
 class GovernBlueprintVersion(object):
     """
     A handle to interact with a blueprint version on the Govern instance.
-    Do not create this directly, use :meth:`dataikuapi.govern.BlueprintDesigner.Blueprint.get_version`
+    Do not create this directly, use :meth:`dataikuapi.govern.blueprint.GovernBlueprint.get_version()`
     """
 
     def __init__(self, client, blueprint_id, blueprint_version_id):
@@ -142,6 +142,7 @@ class GovernBlueprintVersion(object):
     def id(self):
         """
         Return the blueprint version id.
+        :rtype: str
         """
         return self.blueprint_version_id
 
@@ -149,8 +150,8 @@ class GovernBlueprintVersion(object):
         """
         Gets the definition of this blueprint version.
 
-        :returns: The definition of the blueprint version as an object.
-        :rtype: :class:`dataikuapi.govern.models.GovernBlueprintVersionDefinition`
+        :return: The definition of the blueprint version as an object.
+        :rtype: :class:`dataikuapi.govern.blueprint.GovernBlueprintVersionDefinition`
 
         """
         definition = self.client._perform_json(
@@ -173,7 +174,7 @@ class GovernBlueprintVersion(object):
 class GovernBlueprintVersionDefinition(object):
     """
     A handle to interact with a blueprint version definition on the Govern instance.
-    Do not create this directly, use :meth:`dataikuapi.govern.models.GovernBlueprintVersion.get_definition`
+    Do not create this directly, use :meth:`dataikuapi.govern.blueprint.GovernBlueprintVersion.get_definition()`
     """
 
     def __init__(self, client, blueprint_id, blueprint_version_id, definition):
@@ -186,7 +187,7 @@ class GovernBlueprintVersionDefinition(object):
         """
         Get raw definition of the blueprint version.
 
-        :return: the raw definition of blueprint, as a dict.
+        :return: The raw definition of blueprint, as a dict.
         :rtype: dict
         """
         return self.definition
@@ -231,7 +232,7 @@ class GovernBlueprintVersionDefinition(object):
         :param str field_definition_id: id of the desired field
         :rtype: dict
         """
-        return self.definition.get("fieldDefinitions").get(field_definition_id)
+        return self.definition.get("fieldDefinitions", {}).get(field_definition_id)
 
     def get_workflow_definition(self):
         """
